Detumbling update
The magnetic coil characteristics shall be introduced already inversed to calculate the intensities when this matrix were not diagonal. 
This latter calculation is modified to incorporate this case.

The magnetic moment calculation is also corrected: the B_module division must be squared.

The constant gainConstant is also simplified by means of the mean orbital motion.