ADCS review
Eclipse threshold is modified to account for a 20% output of the solar panels out of its maximum. Eclipse_decision function is modified to avoid using the solar sensor data.

Solar ephemerides in solar_determination are now normalized.

Active control updating between iterations is revised: intensities were not saved.

Variables names in onboardcontrol are changed to be consistent all along the system (stateVars)